requirements.txt
requests>=2.31.0
typer>=0.12.0
streamlit>=1.36.0
python-dotenv>=1.0.1
